+++ lib/VWOBuilder.ts
@@ -26,7 +26,7 @@ import { SettingsService } from './services/SettingsService';
 
 import { DebugLogMessagesEnum, InfoLogMessagesEnum } from './enums/log-messages';
 import { IVWOOptions } from './models/VWOOptionsModel';
-import { isNumber } from './utils/DataTypeUtil';
+import { isEmptyObject, isNumber } from './utils/DataTypeUtil';
 import { cloneObject, getFormattedErrorMessage } from './utils/FunctionUtil';
 import { buildMessage } from './utils/LogMessageUtil';
 import { Deferred } from './utils/PromiseUtil';
@@ -87,6 +87,9 @@ export class VWOBuilder implements IVWOBuilder {
    * @returns {this} The instance of this builder.
    */
   setNetworkManager(): this {
+    if (this.options.edgeConfig && !isEmptyObject(this.options?.edgeConfig)) {
+      this.options.shouldWaitForTrackingCalls = true;
+    }
     const networkInstance = NetworkManager.Instance;
     // Attach the network client from options
     networkInstance.attachClient(
@@ -108,7 +111,7 @@ export class VWOBuilder implements IVWOBuilder {
 
   initBatching(): this {
     // If edge config is provided, set the batch event data to the default values
-    if (this.options.edgeConfig && Object.keys(this.options?.edgeConfig).length > 0) {
+    if (this.options.edgeConfig && !isEmptyObject(this.options?.edgeConfig)) {
       const edgeConfigModel = new EdgeConfigModel().modelFromDictionary(this.options.edgeConfig);
       this.options.batchEventData = {
         eventsPerRequest: edgeConfigModel.getMaxEventsToBatch(),
@@ -173,22 +176,18 @@ export class VWOBuilder implements IVWOBuilder {
 
   /**
    * Fetches settings asynchronously, ensuring no parallel fetches.
-   * @param {boolean} [force=false] - Force fetch ignoring cache.
    * @returns {Promise<SettingsModel>} A promise that resolves to the fetched settings.
    */
-  fetchSettings(force?: boolean): Promise<Record<any, any>> {
+  fetchSettings(): Promise<Record<any, any>> {
     const deferredObject = new Deferred();
 
     // Check if a fetch operation is already in progress
     if (!this.isSettingsFetchInProgress) {
       this.isSettingsFetchInProgress = true;
-      this.settingFileManager.getSettings(force).then((settings: Record<any, any>) => {
+      this.settingFileManager.getSettings().then((settings: Record<any, any>) => {
         this.isSettingsValid = this.settingFileManager.isSettingsValid;
         this.settingsFetchTime = this.settingFileManager.settingsFetchTime;
-        // if force is false, update original settings, if true the request is from polling and no need to update original settings
-        if (!force) {
-          this.originalSettings = settings;
-        }
+        this.originalSettings = settings;
 
         this.isSettingsFetchInProgress = false;
         deferredObject.resolve(settings);
@@ -203,20 +202,19 @@ export class VWOBuilder implements IVWOBuilder {
 
   /**
    * Gets the settings, fetching them if not cached or if forced.
-   * @param {boolean} [force=false] - Force fetch ignoring cache.
    * @returns {Promise<SettingsModel>} A promise that resolves to the settings.
    */
-  getSettings(force?: boolean): Promise<Record<any, any>> {
+  getSettings(): Promise<Record<any, any>> {
     const deferredObject = new Deferred();
 
     try {
       // Use cached settings if available and not forced to fetch
-      if (!force && this.settings) {
+      if (this.settings) {
         LogManager.Instance.info('Using already fetched and cached settings');
         deferredObject.resolve(this.settings);
       } else {
-        // Fetch settings if not cached or forced
-        this.fetchSettings(force).then((settings: Record<any, any>) => {
+        // Fetch settings if not cached
+        this.fetchSettings().then((settings: Record<any, any>) => {
           deferredObject.resolve(settings);
         });
       }
@@ -226,7 +224,7 @@ export class VWOBuilder implements IVWOBuilder {
         {
           err: getFormattedErrorMessage(err),
         },
-        { an: force ? Constants.POLLING : ApiEnum.INIT },
+        { an: ApiEnum.INIT },
         false,
       );
       deferredObject.resolve({});
@@ -242,6 +240,7 @@ export class VWOBuilder implements IVWOBuilder {
     if (this.options.storage) {
       // Attach the storage connector from options
       this.storage = Storage.Instance.attachConnector(this.options.storage);
+      this.settingFileManager.isStorageServiceProvided = true;
     } else if (typeof process === 'undefined' && typeof window !== 'undefined' && window.localStorage) {
       // eslint-disable-next-line @typescript-eslint/no-var-requires
       const { BrowserStorageConnector } = require('./packages/storage/connectors/BrowserStorageConnector');
@@ -258,6 +257,7 @@ export class VWOBuilder implements IVWOBuilder {
           service: this.options?.clientStorage?.provider === sessionStorage ? `Session Storage` : `Local Storage`,
         }),
       );
+      this.settingFileManager.isStorageServiceProvided = true;
     } else {
       // Set storage to null if no storage options provided
       this.storage = null;
@@ -424,7 +424,7 @@ export class VWOBuilder implements IVWOBuilder {
   checkAndPoll(): void {
     const poll = async () => {
       try {
-        const latestSettings = await this.getSettings(true);
+        const latestSettings = await this.getSettings();
         if (
           latestSettings &&
           Object.keys(latestSettings).length > 0 &&