+++ lib/services/SettingsService.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 import { dynamic } from '../types/Common';
-import { Storage } from '../packages/storage';
 import { LogManager } from '../packages/logger';
 import { NetworkManager, RequestModel, ResponseModel } from '../packages/network-layer';
 
@@ -30,6 +29,8 @@ import { createNetWorkAndRetryDebugEvent, getSettingsPath } from '../utils/Netwo
 import { sendDebugEventToVWO } from '../utils/DebuggerServiceUtil';
 import { getFormattedErrorMessage } from '../utils/FunctionUtil';
 import { ApiEnum } from '../enums/ApiEnum';
+import { StorageService } from './StorageService';
+import { isEmptyObject } from '../utils/DataTypeUtil';
 
 interface ISettingsService {
   sdkKey: string;
@@ -52,6 +53,8 @@ export class SettingsService implements ISettingsService {
   private static instance: SettingsService;
   isSettingsValid: boolean = false;
   proxyProvided: boolean = false;
+  isStorageServiceProvided: boolean = false;
+  isEdgeEnvironment: boolean = false;
   gatewayServiceConfig: {
     hostname: string | null;
     protocol: string | null;
@@ -67,7 +70,10 @@ export class SettingsService implements ISettingsService {
     this.accountId = options.accountId;
     this.expiry = options?.settings?.expiry || Constants.SETTINGS_EXPIRY;
     this.networkTimeout = options?.settings?.timeout || Constants.SETTINGS_TIMEOUT;
-
+    this.isStorageServiceProvided = options?.isStorageServiceProvided || false;
+    if (options?.edgeConfig && Object.keys(options?.edgeConfig).length > 0) {
+      this.isEdgeEnvironment = true;
+    }
     // if sdk is running in browser environment then set isGatewayServiceProvided to true
     // when gatewayService is not provided then we dont update the url and let it point to dacdn by default
     // Check if sdk running in browser and not in edge/serverless environment
@@ -143,18 +149,6 @@ export class SettingsService implements ISettingsService {
     return SettingsService.instance;
   }
 
-  private setSettingsExpiry() {
-    const settingsTimeout = setTimeout(() => {
-      this.fetchSettingsAndCacheInStorage().then(() => {
-        clearTimeout(settingsTimeout);
-        // again set the timer
-        // NOTE: setInterval could be used but it will not consider the time required to fetch settings
-        // This breaks the timer rythm and also sends more call than required
-        this.setSettingsExpiry();
-      });
-    }, this.expiry);
-  }
-
   async normalizeSettings(settings: Record<any, any>): Promise<Record<any, any>> {
     const normalizedSettings = { ...settings };
     if (!normalizedSettings.features || Object.keys(normalizedSettings.features).length === 0) {
@@ -164,75 +158,6 @@ export class SettingsService implements ISettingsService {
       normalizedSettings.campaigns = [];
     }
     return normalizedSettings;
-  }
-
-  private async handleBrowserEnvironment(
-    storageConnector: any,
-    deferredObject: { resolve: (value: any) => void; reject: (reason?: any) => void },
-  ): Promise<void> {
-    try {
-      const cachedSettings = await storageConnector.getSettingsFromStorage(this.sdkKey, this.accountId);
-
-      if (cachedSettings) {
-        LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_FETCH_FROM_CACHE));
-        deferredObject.resolve(cachedSettings);
-      } else {
-        LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_CACHE_MISS));
-        const freshSettings = await this.fetchSettings();
-        const normalizedSettings = await this.normalizeSettings(freshSettings);
-        // set the settings in storage only if settings are valid
-        this.isSettingsValid = new SettingsSchema().isSettingsValid(normalizedSettings);
-        if (this.isSettingsValid) {
-          await storageConnector.setSettingsInStorage(normalizedSettings);
-        }
-        LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_FETCH_SUCCESS));
-        deferredObject.resolve(normalizedSettings);
-      }
-    } catch (error) {
-      LogManager.Instance.errorLog(
-        'ERROR_FETCHING_SETTINGS',
-        {
-          err: getFormattedErrorMessage(error),
-        },
-        { an: Constants.BROWSER_STORAGE },
-        false,
-      );
-      deferredObject.resolve(null);
-    }
-  }
-
-  private async handleServerEnvironment(deferredObject: {
-    resolve: (value: any) => void;
-    reject: (reason?: any) => void;
-  }): Promise<void> {
-    try {
-      const settings = await this.fetchSettings();
-      const normalizedSettings = await this.normalizeSettings(settings);
-      deferredObject.resolve(normalizedSettings);
-    } catch (error) {
-      LogManager.Instance.errorLog(
-        'ERROR_FETCHING_SETTINGS',
-        {
-          err: getFormattedErrorMessage(error),
-        },
-        { an: ApiEnum.INIT },
-        false,
-      );
-      deferredObject.resolve(null);
-    }
-  }
-
-  private fetchSettingsAndCacheInStorage(): Promise<Record<any, any>> {
-    const deferredObject = new Deferred();
-    const storageConnector = Storage.Instance.getConnector();
-
-    if (typeof process === 'undefined' && typeof XMLHttpRequest !== 'undefined') {
-      this.handleBrowserEnvironment(storageConnector, deferredObject);
-    } else {
-      this.handleServerEnvironment(deferredObject);
-    }
-
-    return deferredObject.promise;
   }
 
   fetchSettings(isViaWebhook = false, apiName = ApiEnum.INIT): Promise<Record<any, any>> {
@@ -324,51 +249,66 @@ export class SettingsService implements ISettingsService {
     }
   }
 
-  getSettings(forceFetch = false): Promise<Record<any, any>> {
+  /**
+   * Gets the settings, fetching them if not cached from storage or server.
+   s* @returns {Promise<Record<any, any>>} A promise that resolves to the settings.
+   */
+  async getSettings(): Promise<Record<any, any>> {
     const deferredObject = new Deferred();
-
-    if (forceFetch) {
-      this.fetchSettingsAndCacheInStorage().then((settings: Record<any, any>) => {
-        deferredObject.resolve(settings);
-      });
-    } else {
-      // const storageConnector = Storage.Instance.getConnector();
-
-      // if (storageConnector) {
-      //   storageConnector
-      //     .get(Constants.SETTINGS)
-      //     .then((storedSettings: dynamic) => {
-      //       if (!isObject(storedSettings)) {
-      //         this.fetchSettingsAndCacheInStorage().then((fetchedSettings) => {
-      //           const isSettingsValid = new SettingsSchema().isSettingsValid(fetchedSettings);
-      //           if (isSettingsValid) {
-      //             deferredObject.resolve(fetchedSettings);
-      //           } else {
-      //             deferredObject.reject(new Error('Settings are not valid. Failed schema validation.'));
-      //           }
-      //         });
-      //       } else {
-      //         deferredObject.resolve(storedSettings);
-      //       }
-      //     })
-      //     .catch(() => {
-      //       this.fetchSettingsAndCacheInStorage().then((fetchedSettings) => {
-      //         deferredObject.resolve(fetchedSettings);
-      //       });
-      //     });
-      // } else {
-      this.fetchSettingsAndCacheInStorage().then((fetchedSettings: Record<any, any>) => {
-        const isSettingsValid = new SettingsSchema().isSettingsValid(fetchedSettings);
-        this.isSettingsValid = isSettingsValid;
+    try {
+      // check if the storage service is provided
+      if (this.isStorageServiceProvided) {
+        const storageService = new StorageService();
+        // get the cached settings from storage
+        const cachedSettings = await storageService.getSettingsFromStorage(
+          this.accountId,
+          this.sdkKey,
+          !this.isEdgeEnvironment,
+        );
+        // if cached settings are found, return the cached settings
+        if (cachedSettings && !isEmptyObject(cachedSettings)) {
+          LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_FETCH_FROM_CACHE));
+          deferredObject.resolve(cachedSettings);
+        } else {
+          // if no cached settings are found, fetch fresh settings from server
+          LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_CACHE_MISS));
+          const freshSettings = await this.fetchSettings();
+          const normalizedSettings = await this.normalizeSettings(freshSettings);
+          // check if the settings are valid
+          this.isSettingsValid = new SettingsSchema().isSettingsValid(normalizedSettings);
+          if (this.isSettingsValid) {
+            // if settings are valid, set the settings in storage
+            LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_FETCH_SUCCESS));
+            await storageService.setSettingsInStorage(this.accountId, this.sdkKey, normalizedSettings);
+            deferredObject.resolve(normalizedSettings);
+          } else {
+            LogManager.Instance.errorLog('INVALID_SETTINGS_SCHEMA', {}, { an: ApiEnum.INIT }, false);
+            deferredObject.resolve({});
+          }
+        }
+      } else {
+        // if the storage service is not provided, fetch fresh settings from server
+        const freshSettings = await this.fetchSettings();
+        const normalizedSettings = await this.normalizeSettings(freshSettings);
+        this.isSettingsValid = new SettingsSchema().isSettingsValid(normalizedSettings);
         if (this.isSettingsValid) {
           LogManager.Instance.info(InfoLogMessagesEnum.SETTINGS_FETCH_SUCCESS);
-          deferredObject.resolve(fetchedSettings);
+          deferredObject.resolve(normalizedSettings);
         } else {
           LogManager.Instance.errorLog('INVALID_SETTINGS_SCHEMA', {}, { an: ApiEnum.INIT }, false);
-
           deferredObject.resolve({});
         }
-      });
+      }
+    } catch (error) {
+      LogManager.Instance.errorLog(
+        'ERROR_FETCHING_SETTINGS',
+        {
+          err: getFormattedErrorMessage(error),
+        },
+        { an: ApiEnum.INIT },
+        false,
+      );
+      deferredObject.resolve({});
     }
 
     return deferredObject.promise;