+++ lib/services/StorageService.ts
@@ -19,9 +19,15 @@ import { Storage } from '../packages/storage';
 import { dynamic } from '../types/Common';
 
 import { LogManager } from '../packages/logger';
-import { isNull, isUndefined } from '../utils/DataTypeUtil';
+import { isEmptyObject, isNull, isUndefined } from '../utils/DataTypeUtil';
 import { Deferred } from '../utils/PromiseUtil';
 import { ApiEnum } from '../enums/ApiEnum';
+import { getFormattedErrorMessage } from '../utils/FunctionUtil';
+import { Constants } from '../constants';
+import { SettingsService } from './SettingsService';
+import { SettingsSchema } from '../models/schemas/SettingsSchemaValidation';
+import { buildMessage } from '../utils/LogMessageUtil';
+import { InfoLogMessagesEnum } from '../enums/log-messages';
 
 export interface IStorageService {
   getDataInStorage(featureKey: any, context: ContextModel): Promise<Record<any, any>>;
@@ -87,5 +93,162 @@ export class StorageService implements IStorageService {
     }
 
     return deferredObject.promise;
+  }
+
+  /**
+   * Gets the settings from storage.
+   * @param accountId The account ID.
+   * @param sdkKey The SDK key.
+   * @returns {Promise<Record<string, any>>} A promise that resolves to the settings or empty object if not found.
+   */
+  async getSettingsFromStorage(
+    accountId: number,
+    sdkKey: string,
+    shouldFetchFreshSettings: boolean = true,
+  ): Promise<Record<string, any>> {
+    const deferredObject = new Deferred();
+    try {
+      const storageInstance = Storage.Instance.getConnector();
+      // check if the storage instance is available and has the getSettings method
+      if (storageInstance && typeof storageInstance.getSettings === 'function') {
+        // get the settingsData from storage
+        const settingsData = await storageInstance.getSettings(accountId, sdkKey);
+        if (!settingsData || isEmptyObject(settingsData)) {
+          // if no settings data is found, resolve the promise with empty object
+          deferredObject.resolve({});
+          return deferredObject.promise;
+        }
+        // if settings data is found, get the settings and last updated timestamp
+        const { settings, timestamp } = settingsData;
+
+        // Check for sdkKey and accountId match
+        if (!settings || settings.sdkKey !== sdkKey || String(settings.accountId ?? settings.a) !== String(accountId)) {
+          LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_CACHE_MISS_KEY_ACCOUNT_ID_MISMATCH));
+          deferredObject.resolve({});
+          return deferredObject.promise;
+        }
+
+        const shouldUseCachedSettings = storageInstance.alwaysUseCachedSettings || Constants.ALWAYS_USE_CACHED_SETTINGS;
+        if (shouldUseCachedSettings) {
+          LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_USING_CACHED_SETTINGS));
+          deferredObject.resolve(settings);
+          return deferredObject.promise;
+        }
+
+        // get the current time
+        const currentTime = Date.now();
+        const settingsTTL = storageInstance.ttl || Constants.SETTINGS_TTL;
+        // check if the settings are expired based on the last updated timestamp
+        if (currentTime - timestamp > settingsTTL) {
+          LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_EXPIRED));
+          deferredObject.resolve({});
+        } else {
+          // if settings are not expired, then return the existing settings and update the settings in storage with new timestamp
+          LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_RETRIEVED_FROM_STORAGE));
+          if (shouldFetchFreshSettings) {
+            // if shouldFetchFreshSettings is true, then fetch fresh settings asynchronously and update the storage with new timestamp
+            this.setFreshSettingsInStorage(accountId, sdkKey);
+          }
+          // decode sdkKey if present in the settings
+          if (settings && settings.sdkKey) {
+            try {
+              settings.sdkKey = atob(settings.sdkKey);
+            } catch (e) {
+              LogManager.Instance.errorLog(
+                'ERROR_DECODING_SDK_KEY_FROM_STORAGE',
+                {
+                  err: getFormattedErrorMessage(e),
+                },
+                { an: Constants.STORAGE },
+              );
+            }
+          }
+          deferredObject.resolve(settings);
+        }
+      } else {
+        deferredObject.resolve({});
+      }
+    } catch (error) {
+      LogManager.Instance.errorLog(
+        'ERROR_READING_SETTINGS_FROM_STORAGE',
+        { err: getFormattedErrorMessage(error) },
+        { an: Constants.STORAGE },
+      );
+      deferredObject.resolve({});
+    }
+
+    return deferredObject.promise;
+  }
+
+  /**
+   * Sets the settings in storage.
+   * @param accountId The account ID.
+   * @param sdkKey The SDK key.
+   * @param settings The settings to be stored.
+   * @returns {Promise<void>} A promise that resolves when the settings are successfully stored.
+   */
+  async setSettingsInStorage(accountId: number, sdkKey: string, settings: Record<string, any>): Promise<void> {
+    const deferredObject = new Deferred();
+    try {
+      const storageInstance = Storage.Instance.getConnector();
+      // check if the storage instance is available and has the setSettings method
+      if (storageInstance && typeof storageInstance.setSettings === 'function') {
+        // clone the settings to avoid mutating the original object
+        const clonedSettings = { ...settings };
+        // create the settings to store object with the cloned settings and the current timestamp
+        const settingsToStore = { settings: clonedSettings, timestamp: Date.now() };
+        // set the settings in storage
+        await storageInstance.setSettings(settingsToStore);
+        LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_SUCCESSFULLY_STORED));
+        deferredObject.resolve();
+      } else {
+        deferredObject.resolve();
+      }
+    } catch (error) {
+      LogManager.Instance.errorLog(
+        'ERROR_STORING_SETTINGS_IN_STORAGE',
+        { err: getFormattedErrorMessage(error) },
+        { an: Constants.STORAGE },
+      );
+      deferredObject.resolve();
+    }
+
+    return deferredObject.promise;
+  }
+
+  /**
+   * Fetches fresh settings and updates the storage with a new timestamp
+   */
+  public async setFreshSettingsInStorage(accountId: number, sdkKey: string): Promise<void> {
+    const deferredObject = new Deferred();
+    // Fetch fresh settings asynchronously and update storage
+    const settingsService = SettingsService.Instance;
+    const storageInstance = Storage.Instance.getConnector();
+    if (settingsService && storageInstance && typeof storageInstance.setSettings === 'function') {
+      settingsService
+        .fetchSettings()
+        .then(async (freshSettings) => {
+          if (freshSettings) {
+            const isSettingsValid = new SettingsSchema().isSettingsValid(freshSettings);
+            if (isSettingsValid) {
+              await this.setSettingsInStorage(accountId, sdkKey, freshSettings);
+              LogManager.Instance.info(buildMessage(InfoLogMessagesEnum.SETTINGS_UPDATED_WITH_FRESH_DATA));
+            }
+          }
+          deferredObject.resolve();
+        })
+        .catch((error) => {
+          LogManager.Instance.errorLog(
+            'ERROR_STORING_FRESH_SETTINGS_IN_STORAGE',
+            {
+              err: getFormattedErrorMessage(error),
+            },
+            { an: Constants.STORAGE },
+          );
+          deferredObject.resolve();
+        });
+
+      return deferredObject.promise;
+    }
   }
 }